feat: product detail 폴더를 feature 하위에 생성
